Split OffloaderController: extract pair-status listener
Extracts the offloader-side pair-status listener cluster into
a new sibling module ``controllers/remote_build/pair_status.py``:
``spawn_pair_status_listener``,
``cancel_pair_status_listener``,
``await_pair_status_flip``, ``apply_pair_status_result``, and
``fire_offloader_pair_status_changed``, plus the
module-private ``_fire_offloader_pair_pin_mismatch`` /
``_fire_offloader_pair_peer_revoked`` helpers (only the apply
branch fires those). Bodies take ``OffloaderController`` as
the first arg; the controller keeps thin bound-method
delegates for the five public entry points because each has
test or cross-module callers:

- ``_spawn_pair_status_listener`` — tests instance-call;
  ``request_pair`` calls it.
- ``_cancel_pair_status_listener`` — tests instance-call;
  ``unpair`` and
  ``peer_link_lifecycle.sweep_stale_pairings_at_endpoint``
  call it.
- ``_await_pair_status_flip`` — tests instance-call directly.
- ``_apply_pair_status_result`` — tests instance-call directly.
- ``_fire_offloader_pair_status_changed`` — ``unpair`` calls
  it (offloader-side path).

The two pin-mismatch / peer-revoked fire helpers stay
module-private (only the apply branch uses them).

Imports moved out of ``offloader.py``: ``await_pair_status``
from ``.peer_link_client``; ``OffloaderPairPeerRevokedData``,
``OffloaderPairStatusChangedData``,
``OffloaderPeerRevokedAlert``, the
``_PAIR_STATUS_RECONNECT_BACKOFF_SECONDS`` constant.

Test repoint: 7 patches in
``tests/test_remote_build_peer_link_client.py`` — 5 sites for
``peer_link_await_pair_status`` and 2 sites for
``_PAIR_STATUS_RECONNECT_BACKOFF_SECONDS`` — moved from the
``rb`` (offloader) namespace to ``rb_pair_status``.

Offloader: 1336 -> 1143 lines (-193). Combined with #766 +
#772 + #774: 1709 -> 1143 (-566, -33%). 379 controller +
peer-link-client tests pass; 676 remote-build tests pass.
